day3/RAG/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
clarifai==11.10.0
clarifai-grpc==11.10.2
clarifai-protocol==0.0.33
click==8.3.0
colorama==0.4.6
contextlib2==21.6.0
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
narwhals==2.11.0
networkx==3.4.2
numpy==2.2.6
openai==2.7.2
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.12.4
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
ruff==0.11.4
safetensors==0.6.2
schema==0.7.5
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.51.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uv==0.7.12
watchdog==6.0.0
yarl==1.22.0
